Don't connect invalid link

+++ sources/communication/links/udp_link.h
@@ -14,11 +14,11 @@ namespace comm
         Q_OBJECT
 
     public:
-        UdpLink(int port = 0, QObject* parent = nullptr);
+        UdpLink(quint16 port = 0, QObject* parent = nullptr);
 
         bool isConnected() const override;
 
-        int port() const;
+        quint16 port() const;
         EndpointList endpoints() const;
         bool autoResponse() const;
 
@@ -29,7 +29,7 @@ namespace comm
         void connectLink() override;
         void disconnectLink() override;
 
-        void setPort(int port);
+        void setPort(quint16 port);
         void addEndpoint(const Endpoint& endpoint);
         void removeEndpoint(const Endpoint& endpoint);
         void clearEndpoints();
@@ -48,7 +48,7 @@ namespace comm
 
     private:
         QUdpSocket* m_socket;
-        int m_port;
+        quint16 m_port;
         EndpointList m_endpoints;
         bool m_autoResponse;
     };
